+++ DGMR-SO/model/generator.py
@@ -159,7 +159,7 @@ class Sampler(snt.Module):
         # activation
         hs = [tf.nn.sigmoid(h) for h in hs]
 
-        return tf.stack(hs, axis=1)
+        return tf.cast(tf.stack(hs, axis=1), tf.float32)
 
 class GBlock(snt.Module):
     """Residual generator block without upsampling."""
